backend/requirements.txt
flask
flask-cors
mysql-connector-python
opencv-python
numpy
easyocr
pillow